verktyg/m_blankett/signal_print/requirements.txt
fpdf2>=2.8.1
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
